finishing it with adding the token to all in tests

+++ migrations/sqls/20220104031711-products-order-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products_order(
   id SERIAL PRIMARY KEY,
-  product_id BIGINT NOT NULL REFERENCES enchanted_stuff(id),
-  order_id BIGINT NOT NULL REFERENCES orders(id),
+  product_id INTEGER NOT NULL REFERENCES enchanted_stuff(id),
+  order_id INTEGER NOT NULL REFERENCES orders(id),
   quantity INTEGER NOT NULL
 );
